Minimal implementation using struct_ops
Implemented struct my_ops, which has only one
member function that can be replaced by a BPF
program. To use struct_ops subsystem, we need to
create an instance of struct bpf_struct_ops and
invoke register_bpf_struct_ops in the init
function. Descriptions are provided as comments
in the source code my_ops.c, so please refer to it.